requirements.txt

### `requirements.txt`


cx_Freeze
numpy
opencv-python
Pillow
pandas
tkinter
